Fix inefficient string concatenation
See FindBugs SBSC_USE_STRINGBUFFER_CONCATENATION

> *SBSC: Method concatenates strings using + in a loop*
The method seems to be building a String using concatenation in a loop. In each iteration, the String is converted to a StringBuffer/StringBuilder, appended to, and converted back to a String. This can lead to a cost quadratic in the number of iterations, as the growing string is recopied in each iteration.