Updates to handle unary postfix operators with support (squashed commits)
As proposed, the classes ```AbstractLazyOperator.java``` and ```AbstractOperator```, and the interfaces ```LazyOperator``` and ```Operator``` were changed back to maintain the original constructors, with a default value <code>unaryOperator = false</code> for the operators, to prevent errors in current implementations. An additional constructor was then created to include the custom <code>boolean</code> parameter <code>unaryOperator</code>, which defines whether or not the operator is unary (if number of expected operands equals 1, <code>unaryOperator = true</code>). This information is later used to allow unary postfix operators to be added and used in the expression, where such information regarding if the operator is unary being retrieved using the method <code>isUnaryOperator</code>. The class ```Expression.java``` updates allow unary postfix operators by adding conditions to handle operand exceptions and TokenType classification using the new information. The files ```TestExposedComponents.java```, ```TestNested.java```, ```TestTokenizer.java``` and ```TestCustoms.java``` were changed back to their original state in the master repository (as of May 25, 2021), with a new test for the custom postfix unary operator being added to the last file. The information was also added to the ```README.md``` file.